// src/app/calendar/interfaces/appointmentInfo.interface.ts
export interface CalendarFormInfo {
  branch:               string;
  service:              string;
  date:                 string;
  time:                 string;
  notes:                string;
  firstname:            string;
  lastname:             string;
  email:                string;
  phone:                string;
  externalId:           string;
  publicId:             string | null;
  terminosCondiciones:  string;
}